# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AdaptationSummary": {
        "additionalProperties": false,
        "description": "Why this screen looks the way it does.\n\nA first-class part of the response, not a string the UI assembles. The\nreason renders verbatim under \"Why this path\", and `topic_id` MUST name a\ntopic present in the roadmap block of the same response — the UI highlights\nthat row to show which evidence the decision rested on.",
        "properties": {
          "reason": {
            "description": "Learner-facing. 'Back to arithmetic first — that is what linear equations depends on.'",
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "state": {
            "enum": [
              "worked_transfer",
              "guided_practice",
              "focused_practice",
              "timed_drill",
              "prerequisite_reset"
            ],
            "title": "State",
            "type": "string"
          },
          "topic_id": {
            "title": "Topic Id",
            "type": "string"
          }
        },
        "required": [
          "state",
          "topic_id",
          "reason"
        ],
        "title": "AdaptationSummary",
        "type": "object"
      },
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiError",
        "type": "object"
      },
      "ApiErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "Stable, machine-readable. e.g. 'question_not_found'.",
            "title": "Code",
            "type": "string"
          },
          "message": {
            "description": "Learner-facing. States what happened and what to do.",
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "description": "Whether trying the same request again could plausibly succeed.",
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ApiErrorBody",
        "type": "object"
      },
      "AskRequest": {
        "additionalProperties": false,
        "properties": {
          "profile": {
            "enum": [
              "rusty",
              "time_poor",
              "hands_free",
              "strong"
            ],
            "title": "Profile",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "profile",
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "AskResponse": {
        "additionalProperties": false,
        "properties": {
          "adaptation": {
            "$ref": "#/components/schemas/AdaptationSummary"
          },
          "blocks": {
            "items": {
              "discriminator": {
                "mapping": {
                  "audio_explainer": "#/components/schemas/AudioExplainerBlock",
                  "explainer_card": "#/components/schemas/ExplainerCardBlock",
                  "flashcards": "#/components/schemas/FlashcardsBlock",
                  "progress_panel": "#/components/schemas/ProgressPanelBlock",
                  "quiz": "#/components/schemas/QuizBlock",
                  "roadmap": "#/components/schemas/RoadmapBlock"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ExplainerCardBlock"
                },
                {
                  "$ref": "#/components/schemas/AudioExplainerBlock"
                },
                {
                  "$ref": "#/components/schemas/QuizBlock"
                },
                {
                  "$ref": "#/components/schemas/RoadmapBlock"
                },
                {
                  "$ref": "#/components/schemas/FlashcardsBlock"
                },
                {
                  "$ref": "#/components/schemas/ProgressPanelBlock"
                }
              ]
            },
            "maxItems": 5,
            "minItems": 1,
            "title": "Blocks",
            "type": "array"
          },
          "cached": {
            "description": "Dev overlay only. Never shown to a learner.",
            "title": "Cached",
            "type": "boolean"
          },
          "spec_id": {
            "title": "Spec Id",
            "type": "string"
          }
        },
        "required": [
          "spec_id",
          "cached",
          "adaptation",
          "blocks"
        ],
        "title": "AskResponse",
        "type": "object"
      },
      "AudioExplainerBlock": {
        "additionalProperties": false,
        "properties": {
          "audio_url": {
            "description": "Must be playable when /ask responds. The frontend never calls TTS.",
            "title": "Audio Url",
            "type": "string"
          },
          "duration_ms": {
            "minimum": 0.0,
            "title": "Duration Ms",
            "type": "integer"
          },
          "script": {
            "title": "Script",
            "type": "string"
          },
          "transcript_shown": {
            "title": "Transcript Shown",
            "type": "boolean"
          },
          "type": {
            "const": "audio_explainer",
            "default": "audio_explainer",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "script",
          "audio_url",
          "duration_ms",
          "transcript_shown"
        ],
        "title": "AudioExplainerBlock",
        "type": "object"
      },
      "Citation": {
        "additionalProperties": false,
        "properties": {
          "chunk_id": {
            "title": "Chunk Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "chunk_id",
          "label"
        ],
        "title": "Citation",
        "type": "object"
      },
      "ExplainerCardBlock": {
        "additionalProperties": false,
        "properties": {
          "citation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Citation"
              },
              {
                "type": "null"
              }
            ]
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/ExplainerStep"
            },
            "maxItems": 5,
            "minItems": 2,
            "title": "Steps",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "const": "explainer_card",
            "default": "explainer_card",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "steps"
        ],
        "title": "ExplainerCardBlock",
        "type": "object"
      },
      "ExplainerStep": {
        "additionalProperties": false,
        "properties": {
          "expr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "TeX, marked at ingestion. The frontend never derives this from prose.",
            "title": "Expr"
          },
          "faded": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Withhold this step's result and invite the learner to supply it. Backward fading: set on the LAST step only, and only when the learner has shown they can already do it.",
            "title": "Faded"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "ExplainerStep",
        "type": "object"
      },
      "Flashcard": {
        "additionalProperties": false,
        "properties": {
          "back": {
            "title": "Back",
            "type": "string"
          },
          "front": {
            "title": "Front",
            "type": "string"
          },
          "topic_id": {
            "title": "Topic Id",
            "type": "string"
          }
        },
        "required": [
          "front",
          "back",
          "topic_id"
        ],
        "title": "Flashcard",
        "type": "object"
      },
      "FlashcardsBlock": {
        "additionalProperties": false,
        "properties": {
          "cards": {
            "items": {
              "$ref": "#/components/schemas/Flashcard"
            },
            "minItems": 1,
            "title": "Cards",
            "type": "array"
          },
          "type": {
            "const": "flashcards",
            "default": "flashcards",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "cards"
        ],
        "title": "FlashcardsBlock",
        "type": "object"
      },
      "MasteryDelta": {
        "additionalProperties": false,
        "properties": {
          "after": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "After",
            "type": "number"
          },
          "before": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Before",
            "type": "number"
          },
          "topic_id": {
            "title": "Topic Id",
            "type": "string"
          }
        },
        "required": [
          "topic_id",
          "before",
          "after"
        ],
        "title": "MasteryDelta",
        "type": "object"
      },
      "ProfileRequest": {
        "additionalProperties": false,
        "properties": {
          "profile": {
            "enum": [
              "rusty",
              "time_poor",
              "hands_free",
              "strong"
            ],
            "title": "Profile",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "profile"
        ],
        "title": "ProfileRequest",
        "type": "object"
      },
      "ProgressPanelBlock": {
        "additionalProperties": false,
        "properties": {
          "next_action_label": {
            "title": "Next Action Label",
            "type": "string"
          },
          "streak": {
            "minimum": 0.0,
            "title": "Streak",
            "type": "integer"
          },
          "type": {
            "const": "progress_panel",
            "default": "progress_panel",
            "title": "Type",
            "type": "string"
          },
          "weakest_topics": {
            "items": {
              "$ref": "#/components/schemas/WeakTopic"
            },
            "maxItems": 3,
            "minItems": 1,
            "title": "Weakest Topics",
            "type": "array"
          }
        },
        "required": [
          "weakest_topics",
          "streak",
          "next_action_label"
        ],
        "title": "ProgressPanelBlock",
        "type": "object"
      },
      "QuizBlock": {
        "additionalProperties": false,
        "properties": {
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuizQuestion"
            },
            "minItems": 1,
            "title": "Questions",
            "type": "array"
          },
          "timer_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Non-null for the `strong` profile only; null for every other.",
            "title": "Timer Seconds"
          },
          "type": {
            "const": "quiz",
            "default": "quiz",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "questions"
        ],
        "title": "QuizBlock",
        "type": "object"
      },
      "QuizQuestion": {
        "additionalProperties": false,
        "properties": {
          "answer_index": {
            "maximum": 4.0,
            "minimum": 0.0,
            "title": "Answer Index",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "maxItems": 5,
            "minItems": 5,
            "title": "Options",
            "type": "array"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "stem": {
            "title": "Stem",
            "type": "string"
          },
          "stem_expr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stem Expr"
          },
          "topic_id": {
            "title": "Topic Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "stem",
          "options",
          "topic_id",
          "answer_index",
          "rationale"
        ],
        "title": "QuizQuestion",
        "type": "object"
      },
      "QuizSubmitRequest": {
        "additionalProperties": false,
        "properties": {
          "elapsed_ms": {
            "description": "Not telemetry. The policy reads this to tell a confident answer from a laboured one, which is what separates a timed drill from guided practice.",
            "minimum": 0.0,
            "title": "Elapsed Ms",
            "type": "integer"
          },
          "question_id": {
            "title": "Question Id",
            "type": "string"
          },
          "selected_index": {
            "minimum": 0.0,
            "title": "Selected Index",
            "type": "integer"
          },
          "topic_id": {
            "title": "Topic Id",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "question_id",
          "topic_id",
          "selected_index",
          "elapsed_ms"
        ],
        "title": "QuizSubmitRequest",
        "type": "object"
      },
      "QuizSubmitResponse": {
        "additionalProperties": false,
        "properties": {
          "answer_index": {
            "maximum": 4.0,
            "minimum": 0.0,
            "title": "Answer Index",
            "type": "integer"
          },
          "correct": {
            "title": "Correct",
            "type": "boolean"
          },
          "mastery": {
            "$ref": "#/components/schemas/MasteryDelta"
          },
          "roadmap_changed": {
            "description": "True when ANYTHING the learner can see on the roadmap changed — ordering, status or evidence claim — not merely the ordering. The client uses this to decide whether to refetch, so a claim-only change must still report true or the screen keeps a stale row.",
            "title": "Roadmap Changed",
            "type": "boolean"
          }
        },
        "required": [
          "correct",
          "answer_index",
          "mastery",
          "roadmap_changed"
        ],
        "title": "QuizSubmitResponse",
        "type": "object"
      },
      "RoadmapBlock": {
        "additionalProperties": false,
        "properties": {
          "steps": {
            "items": {
              "$ref": "#/components/schemas/RoadmapStep"
            },
            "minItems": 1,
            "title": "Steps",
            "type": "array"
          },
          "type": {
            "const": "roadmap",
            "default": "roadmap",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "steps"
        ],
        "title": "RoadmapBlock",
        "type": "object"
      },
      "RoadmapStep": {
        "additionalProperties": false,
        "properties": {
          "evidence_claim": {
            "description": "Learner-facing, never a percentage. Rendered instead of mastery.",
            "minLength": 1,
            "title": "Evidence Claim",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "mastery": {
            "description": "Ordering weight only. Never rendered — the UI shows evidence_claim.",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Mastery",
            "type": "number"
          },
          "status": {
            "enum": [
              "next",
              "done",
              "later"
            ],
            "title": "Status",
            "type": "string"
          },
          "topic_id": {
            "title": "Topic Id",
            "type": "string"
          }
        },
        "required": [
          "topic_id",
          "label",
          "mastery",
          "evidence_claim",
          "status"
        ],
        "title": "RoadmapStep",
        "type": "object"
      },
      "WeakTopic": {
        "additionalProperties": false,
        "properties": {
          "evidence_claim": {
            "minLength": 1,
            "title": "Evidence Claim",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "mastery": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Mastery",
            "type": "number"
          },
          "topic_id": {
            "title": "Topic Id",
            "type": "string"
          }
        },
        "required": [
          "topic_id",
          "label",
          "mastery",
          "evidence_claim"
        ],
        "title": "WeakTopic",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Adaptive GRE/GMAT quantitative reasoning. Composes a screen from observed learner behaviour and explains every instructional choice.",
    "title": "Ecstacy",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/ask": {
      "post": {
        "description": "Compose the next screen from observed behaviour.\n\nMUST BE SIDE-EFFECT FREE. The frontend prefetches the other three profiles\nthe instant the first response lands, so this is called four times per\nquestion, concurrently. It must not persist the profile (four racing writes)\nand question selection must be deterministic for a given evidence state, or\nflipping profiles silently changes the question on screen.",
        "operationId": "ask_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Ask",
        "tags": [
          "learning"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Health",
        "tags": [
          "ops"
        ]
      }
    },
    "/profile": {
      "post": {
        "description": "Persist the delivery preference.\n\n204 with a genuinely empty body. All three of status_code, response_class\nand response_model=None are required, or FastAPI documents a 200 JSON\nresponse and the contract lies about what this returns.",
        "operationId": "set_profile_profile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Set Profile",
        "tags": [
          "learning"
        ]
      }
    },
    "/progress": {
      "get": {
        "operationId": "progress_progress_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "default": "demo",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressPanelBlock"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Progress",
        "tags": [
          "learning"
        ]
      }
    },
    "/quiz/submit": {
      "post": {
        "description": "Record an answer. The only endpoint that moves mastery.",
        "operationId": "quiz_submit_quiz_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizSubmitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Quiz Submit",
        "tags": [
          "learning"
        ]
      }
    },
    "/roadmap": {
      "get": {
        "operationId": "roadmap_roadmap_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "default": "demo",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoadmapBlock"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Roadmap",
        "tags": [
          "learning"
        ]
      }
    }
  }
}